+++ src/template/webpack-ts.js
@@ -9,15 +9,18 @@ const outputDirectory = "dist";
 const webpackMode =
   process.env.NODE_ENV === "production" ? "production" : "development";
 
-console.log(`Webpack Mode === ${chalk.yellowBright(webpackMode)}\n`);
+console.log(`Webpack Mode === ${chalk.blueBright(webpackMode)}\n`);
 
 console.log(
-  `Project Version === ${chalk.yellowBright(process.env.npm_package_version)}`
+  `Project Version === ${chalk.blueBright(process.env.npm_package_version)}`
 );
+
+console.log("");
 
 module.exports = {
   mode: webpackMode,
   devServer: {
+    clientLogLevel: "error",
     contentBase: path.join(__dirname, "public"),
     historyApiFallback: true,
     open: true,