+++ website/src/server/app.ts
@@ -59,10 +59,10 @@ class App {
   public listen(): void {
     this.app.listen(this.port, () => {
       // eslint-disable-next-line no-console
-      console.log(`Mode: ${chalk.yellowBright(process.env.NODE_ENV)}\n`);
+      console.log(`Mode: ${chalk.blueBright(process.env.NODE_ENV)}\n`);
       // eslint-disable-next-line no-console
       console.log(
-        `Server is listening on port: ${chalk.yellowBright(this.port)}\n`,
+        `Server is listening on port: ${chalk.blueBright(this.port)}\n`,
       );
     });
   }